trading/requirements.txt
flask>=3.0
ccxt>=4.0
pandas>=2.0
ta-lib-bin>=0.5;python_version>="3.10"
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=22.0
numpy>=1.24
